requirements.txt
Flask
Gunicorn
certifi
chardet
click
future
idna
itsdangerous
jinja2
markupsafe
python-telegram-bot==10.1.0
requests
urllib3
werkzeug
openai